chore: move type definitions from wayfinding to sdk package
